rdgenai-api-sample/requirements.txt
pandas>=1.5.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
konlpy>=0.6.0
chromadb>=0.4.0
requests>=2.28.0
pycryptodome>=3.17.0
numpy>=1.21.0
scikit-learn>=1.1.0
google-generativeai>=0.3.0 